refactor: move canvas components to Canvas folder
