+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Toaster } from "react-hot-toast"
 import { Provider } from "react-redux"
 import store from "@/redux/store"
 
-import Canvas from "@/components/Canvas"
+import Canvas from "@/components/Canvas/Canvas"
 import EditingPanel from "@/components/EditingPanel"
 
 function App() {
