+++ src/components/Canvas/CanvasFooter.tsx
